roachtest: move perturbation/* tests to their own directory
As part of the effort to make perturbation tests more first class and
owned by separate teams, this commit moves them to their own directory.

Epic: none

Release note: None
